refactor(pipeline): split constants & utilites into common.py
